move corpusdir and newarticledir into package __init__.py, change imports
