+++ allofplos/article_class.py
@@ -5,11 +5,13 @@ import subprocess
 import lxml.etree as et
 import requests
 
+from . import corpusdir
+
 from .transformations import (filename_to_doi, EXT_URL_TMP, INT_URL_TMP,
-                                       BASE_URL_ARTICLE_LANDING_PAGE)
-from .plos_regex import (validate_doi, corpusdir)
+                              BASE_URL_ARTICLE_LANDING_PAGE)
+from .plos_regex import validate_doi
 from .article_elements import (parse_article_date, get_contrib_info,
-                                        match_contribs_to_dicts)
+                               match_contribs_to_dicts)
 
 
 class Article():
