+++ allofplos/plos_corpus.py
@@ -34,7 +34,9 @@ import progressbar
 import requests
 from tqdm import tqdm
 
-from .plos_regex import (validate_doi, corpusdir, newarticledir)
+from . import corpusdir, newarticledir
+
+from .plos_regex import validate_doi
 from .transformations import (BASE_URL_API, EXT_URL_TMP, INT_URL_TMP, URL_TMP, filename_to_doi,
                                        doi_to_path)
